[examples] Chore: Replace node-state-style with the dev-examples version
Since 0.45.0 was released, the DOMImportExtension-based node-state-style
example no longer needs to live in dev-examples (which builds against the
workspace source). Move its code and dependencies into the standalone
examples/node-state-style, which builds against the published 0.45.0
packages, and remove the dev-examples copy.

- Bring over App.tsx, ToolbarPlugin.tsx, and styleState.ts (the
  extension/signal-based toolbar and the DOMImportExtension-based style
  import), dropping the dev- prefix from the internal extension names.
- Prefer the dev-examples dependency versions (shiki 4, lucide-react
  1.17, vite 8, typescript 6, @vitejs/plugin-react 6, cross-env 10,
  rollup 4.61); keep the lexical packages pinned to the published 0.45.0.
- Keep the standalone examples vite/tsconfig setup (vite.config.ts +
  vite.config.monorepo.ts, self-contained tsconfig.json) since the
  example is no longer part of the workspace.
- Mention the DOMImportExtension in the README, and drop the
  dev-examples/node-state-style entry from pnpm-lock.yaml.
