--- dev-examples/node-state-style/index.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Lexical Node State Example</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script src="/src/main.tsx" type="module"></script>
-  </body>
-</html>
