--- dev-examples/node-state-style/src/main.tsx
@@ -1,22 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-import './styles.css';
-
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-
-import App from './App.tsx';
-
-ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <div className="App">
-      <h1>NodeState Style Example</h1>
-      <App />
-    </div>
-  </React.StrictMode>,
-);
