--- dev-examples/node-state-style/src/plugins/ShikiViewPlugin.css
@@ -1,5 +1,0 @@
-.shiki-view-plugin > pre.shiki {
-  margin: 0;
-  padding: 10px;
-  white-space: pre-wrap;
-}
