--- dev-examples/node-state-style/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "useDefineForClassFields": true,
-    "skipLibCheck": true,
-
-    /* Bundler mode */
-    "moduleResolution": "bundler",
-    "allowImportingTsExtensions": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true,
-    "jsx": "react-jsx",
-
-    /* Linting */
-    "strict": true,
-    "noFallthroughCasesInSwitch": true,
-
-    /* Inherit lib + paths from the workspace root so reaching into
-     * workspace package source picks up node globals (`process`,
-     * `global`), `Symbol.dispose`, `findLastIndex`, etc. */
-    "types": ["node"]
-  },
-  "include": ["src", "../../libdefs/globals.d.ts"],
-  "references": [{"path": "./tsconfig.node.json"}]
-}
