+++ examples/node-state-style/README.md
@@ -1,7 +1,9 @@
 # Node State Style example
 
 Here we have an example that demonstrates how NodeState can be used with a
-DOMRenderExtension to override create and export behavior of any node.
+DOMRenderExtension to override the create and export behavior of any node, and a
+DOMImportExtension to capture arbitrary inline `style` properties when importing
+HTML.
 
 **Run it locally:** `pnpm i && pnpm run dev`
