+++ examples/node-state-style/package.json
@@ -18,29 +18,29 @@
     "@lexical/rich-text": "0.45.0",
     "@lexical/selection": "0.45.0",
     "@lexical/utils": "0.45.0",
-    "@shikijs/langs": "^3.3.0",
-    "@shikijs/themes": "^3.3.0",
+    "@shikijs/langs": "^4.2.0",
+    "@shikijs/themes": "^4.2.0",
     "@zag-js/combobox": "^1.9.2",
     "lexical": "0.45.0",
-    "lucide-react": "^0.503.0",
+    "lucide-react": "^1.17.0",
     "prettier": "^3.5.3",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "shiki": "^3.3.0"
+    "shiki": "^4.2.0"
   },
   "devDependencies": {
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.2",
-    "@vitejs/plugin-react": "^5.0.2",
-    "cross-env": "^7.0.3",
+    "@vitejs/plugin-react": "^6.0.2",
+    "cross-env": "^10.1.0",
     "csstype": "^3.1.3",
-    "typescript": "^5.9.2",
-    "vite": "^7.3.2"
+    "typescript": "^6.0.3",
+    "vite": "^8.0.16"
   },
   "optionalDependencies": {
-    "@rollup/rollup-linux-x64-gnu": "4.52.0",
-    "@rollup/rollup-darwin-arm64": "4.52.0",
-    "@rollup/rollup-win32-x64-msvc": "4.52.0",
-    "@rollup/wasm-node": "4.52.0"
+    "@rollup/rollup-linux-x64-gnu": "4.61.0",
+    "@rollup/rollup-darwin-arm64": "4.61.0",
+    "@rollup/rollup-win32-x64-msvc": "4.61.0",
+    "@rollup/wasm-node": "4.61.0"
   }
 }
