+++ examples/node-state-style/src/App.tsx
@@ -12,12 +12,12 @@ import {HistoryExtension} from '@lexical/history';
 import {ContentEditable} from '@lexical/react/LexicalContentEditable';
 import {LexicalExtensionComposer} from '@lexical/react/LexicalExtensionComposer';
 import {RichTextExtension} from '@lexical/rich-text';
-import {defineExtension, ParagraphNode, TextNode} from 'lexical';
+import {defineExtension} from 'lexical';
 
 import ExampleTheme from './ExampleTheme';
 import {ShikiViewPlugin} from './plugins/ShikiViewPlugin';
 import {StyleViewPlugin} from './plugins/StyleViewPlugin';
-import {ToolbarPlugin} from './plugins/ToolbarPlugin';
+import {ToolbarExtension, ToolbarPlugin} from './plugins/ToolbarPlugin';
 import {StyleStateExtension} from './styleState';
 
 const placeholder = 'Enter some rich text...';
@@ -28,13 +28,10 @@ const editorExtension = defineExtension({
     HistoryExtension,
     AutoFocusExtension,
     StyleStateExtension,
+    ToolbarExtension,
   ],
   name: '@lexical/examples/node-state-style',
   namespace: 'NodeState Demo',
-  nodes: [ParagraphNode, TextNode],
-  onError(error: Error) {
-    throw error;
-  },
   theme: ExampleTheme,
 });
