+++ examples/node-state-style/src/plugins/ToolbarPlugin.tsx
@@ -5,20 +5,21 @@
  * LICENSE file in the root directory of this source tree.
  *
  */
+import {batch, signal} from '@lexical/extension';
+import {HistoryExtension} from '@lexical/history';
 import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
+import {useExtensionSignalValue} from '@lexical/react/useExtensionSignalValue';
 import {mergeRegister} from '@lexical/utils';
 import {
   $getSelection,
   $isRangeSelection,
-  CAN_REDO_COMMAND,
-  CAN_UNDO_COMMAND,
   COMMAND_PRIORITY_LOW,
+  defineExtension,
   FORMAT_TEXT_COMMAND,
   REDO_COMMAND,
   SELECTION_CHANGE_COMMAND,
   UNDO_COMMAND,
 } from 'lexical';
-import {useCallback, useEffect, useRef, useState} from 'react';
 
 import {
   $selectionHasStyle,
@@ -30,119 +31,119 @@ function Divider() {
   return <div className="divider" />;
 }
 
-export function ToolbarPlugin() {
-  const [editor] = useLexicalComposerContext();
-  const toolbarRef = useRef(null);
-  const [canUndo, setCanUndo] = useState(false);
-  const [canRedo, setCanRedo] = useState(false);
-  const [isBold, setIsBold] = useState(false);
-  const [isItalic, setIsItalic] = useState(false);
-  const [isUnderline, setIsUnderline] = useState(false);
-  const [isStrikethrough, setIsStrikethrough] = useState(false);
-  const [isStyled, setIsStyled] = useState(false);
-
-  const $updateToolbar = useCallback(() => {
-    const selection = $getSelection();
-    setIsStyled($selectionHasStyle());
-    if ($isRangeSelection(selection)) {
-      // Update text format
-      setIsBold(selection.hasFormat('bold'));
-      setIsItalic(selection.hasFormat('italic'));
-      setIsUnderline(selection.hasFormat('underline'));
-      setIsStrikethrough(selection.hasFormat('strikethrough'));
-    }
-  }, []);
-
-  useEffect(() => {
+/**
+ * Owns the toolbar's reactive state as a handful of signals and keeps
+ * them in sync with the editor. The React `ToolbarPlugin` reads via
+ * {@link useExtensionSignalValue} — no local React state. `canUndo` /
+ * `canRedo` come straight from `HistoryExtension` so this extension
+ * only owns the format / style flags.
+ */
+export const ToolbarExtension = defineExtension({
+  build() {
+    return {
+      isBold: signal(false),
+      isItalic: signal(false),
+      isStrikethrough: signal(false),
+      isStyled: signal(false),
+      isUnderline: signal(false),
+    };
+  },
+  dependencies: [HistoryExtension],
+  name: '@lexical/examples/node-state-style/Toolbar',
+  register(editor, _config, state) {
+    const out = state.getOutput();
+    const $sync = () => {
+      const selection = $getSelection();
+      batch(() => {
+        out.isStyled.value = $selectionHasStyle();
+        if ($isRangeSelection(selection)) {
+          out.isBold.value = selection.hasFormat('bold');
+          out.isItalic.value = selection.hasFormat('italic');
+          out.isUnderline.value = selection.hasFormat('underline');
+          out.isStrikethrough.value = selection.hasFormat('strikethrough');
+        }
+      });
+    };
     return mergeRegister(
-      editor.registerUpdateListener(({editorState}) => {
-        editorState.read(() => {
-          $updateToolbar();
-        });
-      }),
+      editor.registerUpdateListener(({editorState}) => editorState.read($sync)),
       editor.registerCommand(
         SELECTION_CHANGE_COMMAND,
-        (_payload, _newEditor) => {
-          $updateToolbar();
-          return false;
-        },
-        COMMAND_PRIORITY_LOW,
-      ),
-      editor.registerCommand(
-        CAN_UNDO_COMMAND,
-        payload => {
-          setCanUndo(payload);
-          return false;
-        },
-        COMMAND_PRIORITY_LOW,
-      ),
-      editor.registerCommand(
-        CAN_REDO_COMMAND,
-        payload => {
-          setCanRedo(payload);
+        () => {
+          $sync();
           return false;
         },
         COMMAND_PRIORITY_LOW,
       ),
     );
-  }, [editor, $updateToolbar]);
+  },
+});
+
+export function ToolbarPlugin() {
+  const [editor] = useLexicalComposerContext();
+  const canUndo = useExtensionSignalValue(HistoryExtension, 'canUndo');
+  const canRedo = useExtensionSignalValue(HistoryExtension, 'canRedo');
+  const isBold = useExtensionSignalValue(ToolbarExtension, 'isBold');
+  const isItalic = useExtensionSignalValue(ToolbarExtension, 'isItalic');
+  const isUnderline = useExtensionSignalValue(ToolbarExtension, 'isUnderline');
+  const isStrikethrough = useExtensionSignalValue(
+    ToolbarExtension,
+    'isStrikethrough',
+  );
+  const isStyled = useExtensionSignalValue(ToolbarExtension, 'isStyled');
 
   return (
-    <div className="toolbar" ref={toolbarRef}>
+    <div className="toolbar">
       <button
+        type="button"
         disabled={!canUndo}
-        onClick={() => {
-          editor.dispatchCommand(UNDO_COMMAND, undefined);
-        }}
+        onClick={() => editor.dispatchCommand(UNDO_COMMAND, undefined)}
         className="toolbar-item spaced"
         aria-label="Undo">
         <i className="format undo" />
       </button>
       <button
+        type="button"
         disabled={!canRedo}
-        onClick={() => {
-          editor.dispatchCommand(REDO_COMMAND, undefined);
-        }}
+        onClick={() => editor.dispatchCommand(REDO_COMMAND, undefined)}
         className="toolbar-item"
         aria-label="Redo">
         <i className="format redo" />
       </button>
       <Divider />
       <button
-        onClick={() => {
-          editor.dispatchCommand(FORMAT_TEXT_COMMAND, 'bold');
-        }}
+        type="button"
+        onClick={() => editor.dispatchCommand(FORMAT_TEXT_COMMAND, 'bold')}
         className={'toolbar-item spaced ' + (isBold ? 'active' : '')}
         aria-label="Format Bold">
         <i className="format bold" />
       </button>
       <button
-        onClick={() => {
-          editor.dispatchCommand(FORMAT_TEXT_COMMAND, 'italic');
-        }}
+        type="button"
+        onClick={() => editor.dispatchCommand(FORMAT_TEXT_COMMAND, 'italic')}
         className={'toolbar-item spaced ' + (isItalic ? 'active' : '')}
         aria-label="Format Italics">
         <i className="format italic" />
       </button>
       <button
-        onClick={() => {
-          editor.dispatchCommand(FORMAT_TEXT_COMMAND, 'underline');
-        }}
+        type="button"
+        onClick={() => editor.dispatchCommand(FORMAT_TEXT_COMMAND, 'underline')}
         className={'toolbar-item spaced ' + (isUnderline ? 'active' : '')}
         aria-label="Format Underline">
         <i className="format underline" />
       </button>
       <button
-        onClick={() => {
-          editor.dispatchCommand(FORMAT_TEXT_COMMAND, 'strikethrough');
-        }}
+        type="button"
+        onClick={() =>
+          editor.dispatchCommand(FORMAT_TEXT_COMMAND, 'strikethrough')
+        }
         className={'toolbar-item spaced ' + (isStrikethrough ? 'active' : '')}
         aria-label="Format Strikethrough">
         <i className="format strikethrough" />
       </button>
       <Divider />
       <button
-        onClick={() => {
+        type="button"
+        onClick={() =>
           editor.dispatchCommand(
             PATCH_TEXT_STYLE_COMMAND,
             isStyled
@@ -151,8 +152,8 @@ export function ToolbarPlugin() {
                   'text-shadow':
                     '1px 1px 2px red, 0 0 1em blue, 0 0 0.2em blue',
                 },
-          );
-        }}
+          )
+        }
         className="toolbar-item spaced"
         aria-label="Toggle Text Style">
         <i className={'text-shadow ' + (isStyled ? 'active' : '')} />
